dispatch structs definition
