+++ cgo/kuzzle/security.go
@@ -11,7 +11,6 @@ import "C"
 
 import (
 	"unsafe"
-
 	"github.com/kuzzleio/sdk-go/kuzzle"
 	"github.com/kuzzleio/sdk-go/types"
 	"github.com/kuzzleio/sdk-go/security"
